+++ pycharm/src/main/java/rw/dialogs/DialogFactory.java
@@ -4,20 +4,6 @@ import com.google.common.annotations.VisibleForTesting;
 import com.intellij.openapi.application.ApplicationManager;
 import com.intellij.openapi.diagnostic.Logger;
 import com.intellij.openapi.project.Project;
-import com.intellij.openapi.ui.DialogWrapper;
-import com.intellij.openapi.util.NlsContexts;
-import com.intellij.ui.UIBundle;
-import com.intellij.ui.components.JBLabel;
-import com.intellij.util.ui.JBFont;
-import com.intellij.util.ui.JBUI;
-import com.intellij.util.ui.components.BorderLayoutPanel;
-import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
-import rw.audit.RwSentry;
-
-import javax.swing.*;
-import java.awt.*;
-import java.lang.reflect.InvocationTargetException;
 
 
 public class DialogFactory {
@@ -45,6 +31,12 @@ public class DialogFactory {
         return dialog.getResult();
     }
 
+    public boolean showSurveyDialog(Project project) {
+        SurveyDialog dialog = new SurveyDialog(project);
+        dialog.show();
+        return dialog.getResult();
+    }
+
     public void showFirstUserErrorDialog(Project project) {
         ApplicationManager.getApplication().invokeLater(() -> {
             FirstUserError dialog = new FirstUserError(project);
@@ -52,7 +44,7 @@ public class DialogFactory {
         });
     }
 
-    public void showFirstFrameErrorDialog(Project project) {
+    public void showFirstThreadErrorDialog(Project project) {
         ApplicationManager.getApplication().invokeLater(() -> {
             FirstFrameError dialog = new FirstFrameError(project);
             dialog.show();