+++ pycharm/src/main/java/rw/handler/runConf/BaseRunConfHandler.java
@@ -6,9 +6,14 @@ import com.intellij.execution.runners.ExecutionEnvironment;
 import com.intellij.execution.ui.RunContentDescriptor;
 import com.intellij.execution.ui.RunContentManager;
 import com.intellij.execution.ui.RunContentWithExecutorListener;
+import com.intellij.execution.ui.RunnerLayoutUi;
 import com.intellij.openapi.Disposable;
 import com.intellij.openapi.project.Project;
+import com.intellij.ui.content.Content;
 import com.intellij.util.messages.MessageBusConnection;
+import com.intellij.xdebugger.XDebuggerManager;
+import com.intellij.xdebugger.frame.XStackFrame;
+import com.intellij.xdebugger.impl.XDebugSessionImpl;
 import com.jetbrains.python.run.AbstractPythonRunConfiguration;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
@@ -16,20 +21,30 @@ import org.jetbrains.annotations.VisibleForTesting;
 import rw.action.RunType;
 import rw.highlights.SolutionHighlightManager;
 import rw.icons.IconPatcher;
-import rw.profile.FrameProgressRenderer;
-import rw.profile.LineProfiler;
+import rw.icons.Icons;
+import rw.profile.*;
+import rw.quickconfig.ProfilerType;
+import rw.quickconfig.QuickConfig;
+import rw.quickconfig.QuickConfigCallback;
+import rw.quickconfig.QuickConfigState;
+import rw.session.cmds.QuickConfigCmd;
+import rw.session.events.Event;
+import rw.session.events.Handshake;
+import rw.session.events.ModuleUpdate;
 import rw.stack.Stack;
 import rw.handler.sdk.SdkHandler;
 import rw.handler.sdk.SdkHandlerFactory;
 import rw.highlights.ErrorHighlightManager;
-import rw.profile.ProfilePreviewRenderer;
 import rw.service.Service;
 import rw.session.Session;
 
 import java.io.File;
 import java.util.HashSet;
-import java.util.List;
+import java.util.Map;
 import java.util.Set;
+
+import static java.util.Map.entry;
+
 
 public abstract class BaseRunConfHandler implements Disposable {
     private final SolutionHighlightManager solutionHighlightManager;
@@ -39,39 +54,66 @@ public abstract class BaseRunConfHandler implements Disposable {
     ExecutionEnvironment executionEnvironment;
     Stack stack;
     FrameProgressRenderer frameProgressRenderer;
-    LineProfiler lineProfiler;
+    TimeProfiler timeProfiler;
+    MemoryProfiler memoryProfiler;
+    LineProfiler activeProfiler;
+    NoneProfiler noneProfiler;
+
     Session session;
     ErrorHighlightManager errorHighlightManager;
-    ProfilePreviewRenderer profilePreviewRenderer;
     Project project;
     MessageBusConnection messageBusConnection;
     Set<File> watchedFiles;
     boolean active;
     @Nullable
     RunType runType;
+    QuickConfig quickConfig;
+    boolean firstActivate;
+    Map<ProfilerType, LineProfiler> profilerTypeToProfiler;
 
     public BaseRunConfHandler(RunConfiguration runConf) {
         this.runConf = (AbstractPythonRunConfiguration<?>) runConf;
         this.project = this.runConf.getProject();
 
+        BaseRunConfHandler This = this;
+        this.quickConfig = new QuickConfig(new QuickConfigCallback() {
+            @Override
+            public void onChange(QuickConfigState state) {
+                This.onQuickConfigChange(state);
+            }
+        });
+
         this.sdkHandler = SdkHandlerFactory.factory(this.runConf.getSdk());
-        this.stack = new Stack(this.project);
+        this.stack = new Stack(this.project, this);
         this.frameProgressRenderer = new FrameProgressRenderer(this.project);
-        this.lineProfiler = new LineProfiler();
+
+        this.timeProfiler = new TimeProfiler(this.project, this.quickConfig);
+        this.memoryProfiler = new MemoryProfiler(this.project, this.quickConfig);
+        this.noneProfiler = new NoneProfiler(this.project, this.quickConfig);
+
         this.session = new Session(this.project, this);
         this.errorHighlightManager = new ErrorHighlightManager(this.project);
         this.solutionHighlightManager = new SolutionHighlightManager(this.project);
-        this.profilePreviewRenderer = new ProfilePreviewRenderer(this.project, this.stack, this.lineProfiler);
 
         this.watchedFiles = new HashSet<>();
         this.active = true;
+        this.firstActivate = true;
+
+        this.profilerTypeToProfiler = Map.ofEntries(
+                entry(ProfilerType.TIME, this.timeProfiler),
+                entry(ProfilerType.MEMORY, this.memoryProfiler),
+                entry(ProfilerType.NONE, this.noneProfiler));
+
+        this.activeProfiler = this.profilerTypeToProfiler.get(this.quickConfig.getState().getProfiler());
 
         this.handleJbEvents();
     }
 
-    abstract public void beforeRun(RunType runType);
+    public void beforeRun(RunType runType) {
+    }
 
-    abstract public void onProcessStarted(RunContentDescriptor descriptor);
+    public void onProcessStarted(RunContentDescriptor descriptor) {
+    }
 
     abstract public void afterRun();
 
@@ -82,12 +124,12 @@ public abstract class BaseRunConfHandler implements Disposable {
     }
 
     public @NotNull
-    String convertPathToLocal(@NotNull String remotePath) {
+    String convertPathToLocal(@NotNull String remotePath, boolean warnMissing) {
         return remotePath;
     }
 
     public @NotNull
-    String convertPathToRemote(@NotNull String localPath) {
+    String convertPathToRemote(@NotNull String localPath, boolean warnMissing) {
         return localPath;
     }
 
@@ -128,7 +170,6 @@ public abstract class BaseRunConfHandler implements Disposable {
 
     private void handleJbEvents() {
         this.messageBusConnection = this.project.getMessageBus().connect(Service.get());
-
         BaseRunConfHandler This = this;
 
         this.messageBusConnection.subscribe(RunContentManager.TOPIC, new RunContentWithExecutorListener() {
@@ -156,20 +197,46 @@ public abstract class BaseRunConfHandler implements Disposable {
 
     public void activate() {
         this.errorHighlightManager.activate();
+        if (this.activeProfiler != null) {
+            this.activeProfiler.activate();
+        }
         this.solutionHighlightManager.activate();
-        this.profilePreviewRenderer.activate();
         this.frameProgressRenderer.activate();
         this.active = true;
         IconPatcher.refresh(this.getProject());
+
+        if (this.firstActivate) {
+            this.onFirstActivate();
+        }
+        this.firstActivate = false;
+    }
+
+    public @Nullable XDebugSessionImpl getDebugSession() {
+        XDebugSessionImpl ret = ((XDebugSessionImpl) XDebuggerManager.getInstance(project).getCurrentSession());
+        return ret;
+    }
+
+    public void onFirstActivate() {
+        XDebugSessionImpl debugSession = this.getDebugSession();
+        assert debugSession != null;
+        String id = Integer.toString(debugSession.hashCode());
+        RunnerLayoutUi layout = RunnerLayoutUi.Factory.getInstance(this.project).create(id, "re_runner", "re_session",
+                this);
+        Content content = layout.createContent(id, quickConfig.getContent(), "loadium", Icons.ProductIcon, null);
+        debugSession.getUI().addContent(content);
     }
 
     public void deactivate() {
         this.errorHighlightManager.deactivate();
         this.solutionHighlightManager.deactivate();
-        this.profilePreviewRenderer.deactivate();
+        this.activeProfiler.deactivate();
         this.frameProgressRenderer.deactivate();
         this.active = false;
         IconPatcher.refresh(this.getProject());
+    }
+
+    public QuickConfig getQuickConfig() {
+        return this.quickConfig;
     }
 
     public void addWatched(Set<File> files) {
@@ -186,21 +253,33 @@ public abstract class BaseRunConfHandler implements Disposable {
         this.messageBusConnection.dispose();
     }
 
-    public ProfilePreviewRenderer getProfilePreviewRenderer() {
-        return profilePreviewRenderer;
-    }
-
     // Test methods ################
     @VisibleForTesting
     public void __setErrorHighlightManager(ErrorHighlightManager errorHighlightManager) {
         this.errorHighlightManager = errorHighlightManager;
     }
 
-    public LineProfiler getTimeProfiler() {
-        return this.lineProfiler;
+    public LineProfiler getActiveProfiler() {
+        return this.activeProfiler;
     }
 
-    public Session getSession() {
+    @NotNull public Session getSession() {
         return this.session;
+    }
+
+    private void onQuickConfigChange(QuickConfigState state) {
+        QuickConfigCmd cmd = new QuickConfigCmd(state);
+        this.getSession().send(cmd);
+        this.activeProfiler.deactivate();
+        this.activeProfiler = this.profilerTypeToProfiler.get(state.getProfiler());
+        this.activeProfiler.activate();
+    }
+
+    public TimeProfiler getTimeProfiler() {
+        return this.timeProfiler;
+    }
+
+    public MemoryProfiler getMemoryProfiler() {
+        return this.memoryProfiler;
     }
 }