+++ pycharm/src/main/java/rw/handler/runConf/RemoteRunConfHandler.java
@@ -1,15 +1,18 @@
 package rw.handler.runConf;
 
+import com.intellij.CommonBundle;
 import com.intellij.execution.configurations.RunConfiguration;
 import com.intellij.execution.process.BaseProcessHandler;
 import com.intellij.execution.ui.RunContentDescriptor;
-import com.intellij.remote.*;
-import com.jetbrains.python.remote.PyRemoteSdkAdditionalDataBase;
+import com.intellij.openapi.application.ApplicationManager;
+import com.intellij.openapi.ui.Messages;
 import org.jetbrains.annotations.NotNull;
 import rw.action.RunType;
 import rw.audit.RwSentry;
 import rw.config.Config;
 import rw.config.ConfigManager;
+
+import java.lang.reflect.Method;
 
 
 public class RemoteRunConfHandler extends PythonRunConfHandler {
@@ -20,11 +23,17 @@ public class RemoteRunConfHandler extends PythonRunConfHandler {
 
     public void onProcessStarted(RunContentDescriptor descriptor) {
         try {
-            RemoteSshProcess process = (RemoteSshProcess)((BaseProcessHandler) descriptor.getProcessHandler()).getProcess();
-            PyRemoteSdkAdditionalDataBase additionalData = ((PyRemoteSdkAdditionalDataBase) this.runConf.getSdk().getSdkAdditionalData());
-            RemoteSdkCredentials credentials = additionalData.getRemoteSdkCredentials(this.runConf.getProject(), true);
+            BaseProcessHandler processHandler = ((BaseProcessHandler) descriptor.getProcessHandler());
+            Object process = processHandler.getProcess();
+            Method addRemoteTunnel = process.getClass().getMethod("addRemoteTunnel", int.class, String.class, int.class);
+            Method getSession = process.getClass().getMethod("getSession");
 
-            process.addRemoteTunnel(this.session.getPort(), credentials.getHost(), this.session.getPort());
+            Object session  = getSession.invoke(process);
+            Method getHost = session.getClass().getMethod("getHost");
+
+            String host = (String) getHost.invoke(session);
+
+            addRemoteTunnel.invoke(process, this.session.getPort(), host, this.session.getPort());
         } catch (Exception e) {
             RwSentry.get().captureException(e);
         }
@@ -42,23 +51,39 @@ public class RemoteRunConfHandler extends PythonRunConfHandler {
 
     @NotNull
     @Override
-    public String convertPathToLocal(@NotNull String remotePath) {
+    public String convertPathToLocal(@NotNull String remotePath, boolean warnMissing) {
         if (this.runConf.getMappingSettings() == null){
+            Messages.showErrorDialog(this.project, "Path mappings are missing", "Missing Path Mappings");
             return remotePath;
         }
 
         String ret = this.runConf.getMappingSettings().convertToLocal(remotePath);
+
+        if(warnMissing && ret.equals(remotePath)) {
+            ApplicationManager.getApplication().invokeLater(() -> {
+                Messages.showErrorDialog(this.project, String.format("Could not convert remote path %s to the local one. " +
+                        "Add path mappings to resolve this", remotePath), "Missing Path Mappings");
+            });
+        }
         return ret;
     }
 
     @NotNull
     @Override
-    public String convertPathToRemote(@NotNull String localPath) {
-        if (this.runConf.getMappingSettings() == null){
+    public String convertPathToRemote(@NotNull String localPath, boolean warnMissing) {
+        if (this.runConf.getMappingSettings() == null) {
+            Messages.showErrorDialog(this.project, "Path mappings are missing", "Missing Path Mappings");
             return localPath;
         }
 
         String ret = this.runConf.getMappingSettings().convertToRemote(localPath);
+
+        if(warnMissing && ret.equals(localPath)) {
+            ApplicationManager.getApplication().invokeLater(() -> {
+                Messages.showErrorDialog(this.project, String.format("Could not convert local path %s to the remote one. " +
+                        "Add path mappings to resolve this", localPath), "Missing Path Mappings");
+            });
+        }
         return ret;
     }
 }