+++ pycharm/src/main/java/rw/preferences/PreferencesForm.java
@@ -1,6 +1,9 @@
 package rw.preferences;
 
 import com.intellij.openapi.ui.ComboBox;
+import rw.quickconfig.CumulateType;
+import rw.quickconfig.ErrorHandlingMode;
+import rw.quickconfig.ProfilerType;
 import rw.util.colormap.ColorMaps;
 
 import javax.swing.*;
@@ -12,12 +15,26 @@ public class PreferencesForm {
     private JCheckBox telemetry;
     private JCheckBox sentry;
     private JCheckBox markReloadable;
+    private JComboBox<ProfilerType> defaultProfiler;
+    private JComboBox<CumulateType> defaultCumulateType;
+    private JComboBox<ErrorHandlingMode> errorHandlingMode;
 
     private void createUIComponents() {
         this.blinkDuration = new JSpinner(new SpinnerNumberModel(0, 0, 2000, 10));
         this.timingColormap = new ComboBox<>();
-
         this.timingColormap.setModel(new DefaultComboBoxModel(ColorMaps.get().getAllImages().toArray()));
+
+        this.defaultProfiler = new ComboBox<>();
+        this.defaultProfiler.setModel(new DefaultComboBoxModel(ProfilerType.getAll()));
+        this.defaultProfiler.setSelectedItem(ProfilerType.DEFAULT);
+
+        this.defaultCumulateType = new ComboBox<>();
+        this.defaultCumulateType.setModel(new DefaultComboBoxModel(CumulateType.getAll()));
+        this.defaultCumulateType.setSelectedItem(CumulateType.DEFAULT);
+
+        this.errorHandlingMode = new ComboBox<>();
+        this.errorHandlingMode.setModel(new DefaultComboBoxModel(ErrorHandlingMode.getAll()));
+        this.errorHandlingMode.setSelectedItem(ErrorHandlingMode.DEFAULT);
     }
 
     public JPanel getMainPanel() {
@@ -31,6 +48,9 @@ public class PreferencesForm {
         state.telemetry = this.telemetry.isSelected();
         state.sentry = this.sentry.isSelected();
         state.markReloadable = this.markReloadable.isSelected();
+        state.defaultProfiler = (ProfilerType) this.defaultProfiler.getModel().getSelectedItem();
+        state.defaultCumulateType = (CumulateType) this.defaultCumulateType.getModel().getSelectedItem();
+        state.defaultErrorHandlingMode = (ErrorHandlingMode) this.errorHandlingMode.getModel().getSelectedItem();
         return state;
     }
 
@@ -40,5 +60,8 @@ public class PreferencesForm {
         this.telemetry.setSelected(state.telemetry);
         this.sentry.setSelected(state.sentry);
         this.markReloadable.setSelected(state.markReloadable);
+        this.defaultProfiler.getModel().setSelectedItem(state.defaultProfiler);
+        this.defaultCumulateType.getModel().setSelectedItem(state.defaultCumulateType);
+        this.errorHandlingMode.getModel().setSelectedItem(state.defaultErrorHandlingMode);
     }
 }