+++ pycharm/src/main/java/rw/session/events/FrameEvent.java
@@ -2,13 +2,12 @@ package rw.session.events;
 
 import com.google.gson.annotations.SerializedName;
 
-abstract public class FrameEvent extends Event {
+abstract public class FrameEvent extends FileEvent {
     @SerializedName("frame_id")
     private Long frameId;
 
-    public Long getFrameId() {
-        return frameId;
-    }
+    @SerializedName("fullname")
+    private String fullname;
 
     @Override
     public void handle() {