+++ pycharm/src/main/java/rw/session/events/StackUpdate.java
@@ -9,17 +9,16 @@ public class StackUpdate extends Event {
     private static final Logger LOGGER = Logger.getInstance(StackUpdate.class);
 
     public static final String ID = "StackUpdate";
-    public static final String VERSION = "0.1.0";
 
-    private Map<Long, List<FrameData>> content;
+    private Map<String, List<FrameData>> content;
 
     @Override
     public void handle() {
         LOGGER.debug("Handling StackUpdate");
 
-        for (List<FrameData> frames : this.content.values()) {
+        for (List<FrameData> frames: this.content.values()) {
             for(FrameData f: frames) {
-                f.localPath = this.handler.convertPathToLocal(f.getPath().toString());
+                f.localPath = this.handler.convertPathToLocal(f.getPath().toString(), false);
             }
         }
 
@@ -27,7 +26,7 @@ public class StackUpdate extends Event {
         this.handler.getStack().onStackUpdateEvent(this);
     }
 
-    public Map<Long, List<FrameData>> getContent() {
+    public Map<String, List<FrameData>> getContent() {
         return content;
     }
 }